requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=2.0.0
pydantic==2.6.1
python-multipart==0.0.6
aiohttp==3.9.3
e2b-code-interpreter>=2.0.0
playwright==1.41.2
beautifulsoup4==4.12.3
requests==2.31.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.21
matplotlib==3.8.2
pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
yfinance==0.2.36
arxiv==2.1.0
python-dotenv==1.0.1
jinja2==3.1.3
weasyprint==61.1
scipy==1.12.0
opencv-python-headless==4.9.0.80
aiofiles==23.2.1
psutil==5.9.8
jupyter_client>=8.0.0
